Adding a new controller and view for the api. The API functions will be api/sources, api/sites, api/methods, api/variables and api/values. The api/sources and api/values are working, but the api/methods, api/variables and api/sites still must be created.
Right now the user must send the username and password in the JSON in the POST request body together with the uploaded values in every API function call.
We'll replace sending the username and password by sending an API key once we create a functionality to store the API key in the database in the users table.