check for invalid IDs with addErrorIfInvalid()
The new function addErrorIfInvalid() checks if an ID is contained in
an array of expected IDs. If not, an error is added and TRUE is
returned, indicating that an error occurred. If the ID is contained,
FALSE is returned. Now, multiple ID checks are performed within a
single if-expression by chaining calls to addErrorIfInvalid() together
with the logical 'or' operator.