control all API errors via loadApiErrorView()
Extract loading of the 'apierror' view with the different API error
messages from the public API methods into a new function
loadApiErrorView(). In this function the list of required parameters
and an example are set depending on the given method. Some methods
can use the same parameter list and example. Common parts of the
error messages appear only once.