improve (fix) JavaScript generation in getData()
The check whether to put a comma BEHIND a tuple was based on
comparing the number of written tuples with the total number
of tuples (including possible 'NoDataValues'). However, the
number of written tuples was only incremented for
non-'NoDataValues'. Now, a comma (and EOL) are written
IN FRONT OF each tuple, except for the first one.