refactor processFiles() with new helper functions
This refactoring increases the readability of processFiles().
New function handleHeaderRow() is called only once for the first row.
It returns the indices for timestamp and data value by reference.
New function handleDataRow() is called for all following rows.
Instead of $flag now $row (now starting with 0 but being incremented
in any case) is used to indicate the header row.