extract configurational parts from datapoint functions
For each high-level function a mapping between long parameter names
(according to the column names in the ODM data model, if applicable)
and short parameter names as given in the URL is defined in
getApiConfiguration(). All high-level functions now use the
new function getAndValidateInputs() to read and rename parameter
values (returned in $input) according to the configuration and to
check whether all parameters are set. Within the functions now
the long parameter names are used which increases the readability.
Function getInputs() needed a change since it is now called with
an assiciative array instead of an indexed array.