base getData(), getDataJSON() and export() on generic
Since the overall structure of getData(), getDataJSON() and export()
is similar, these functions are now based on the new function
getData_generic() that represents this structure. It becomes obvious
that not only the output that these functions generate depends on the
selected method (see new function outputOrExportData()) but also the
way data is gathered from the datapoints model (see new function
getDataFromModel()).