new generic function for adding Datapoints
Since the overall structure of addvalue() and addmultiplevalue()
of the Datapoint class is similar, these functions are now just
wrappers around the new function addvalue_generic() that
implements this structure. The differences between the two methods
become obvious (e.g. there are form validation rules only defined
for 'addvalue', why?). The actual application of the method happens
in the new function applyMethod().