create and apply new class API_Config
Write a new class API_Config providing functions that help to configure
the parameters of the API methods. A parameter is defined by calling
API_Config::parameter(), giving the parameter's name, an example value
and a flag indicating if the parameter is required or not. Use
API_Config::exampleCall() to generate an example call from the
configuration and the method name. Adapt getApiConfiguration() in
the Datapoint controller to now using the methods of API_Config.