read fields to be exported to CSV from config file
A new configuration item 'export_fields' is defined giving the names
of the fields that are to be contained in the exported CSV file. If the
item 'export_fields' is not given in the configuration file, the
program behaves in the same way as before, including only the fields
'ValueID', 'DataValue' and LocalDateTime' in the exported file.