allow for extended export (resolving foreign keys)
If the new argument $extended in getData(), getResultData() or
getQueryForGetData() is set to TRUE (default: FALSE), the database query
is altered in such a way that the tables Sites, Variables, Sources,
Methods are joined and the additional fields SiteCode, VariableCode,
Organization and MethodDescription are queried. The extended export can
be switched on or off by setting the configuration item 'extended_export'
to 'TRUE' or 'FALSE', respectively.