allow to set UTCOffset in CSV import file
Change interface of validateFields() in order to be more flexible when
adding further fields to be validated. If field "UTCOffset" ist given,
check if the value is numeric. In createDataPoint(), fill result data
structure with given fields (that may contain "UTCOffset") first before
calculating DateTimeUTC. Rename $localtimestamp with $localtime for
better readability.