configure the navigation bar entries in an array
For each main menu item provide an element in the array $menuConfig.
Each element is an array with elements 'keyword' (to be looked up in
the language table to get the content of the menu entry), 'id' (given
to the HTML element <ul>) and 'items' (array of HTML snippets
representing the list items). Only with the next commit, this
configuration will be used.