add class Excel inheriting from PHPExcel
Provide two functions:

output_as_xls() gets an array as input and creates an Excel worksheet
"data" containing the content of that array. The first row of the Excel
worksheet is used to show the column headers, taken from the array_keys
of the first row. The content of the Excel file is printed to STDOUT.

read_xls() reads an Excel file from the given path and returns an
array containing the content of the first worksheet. Each array element
represents a row of the worksheet and is itself an array of elements
being named A, B, C, etc. (as the columns are named in Excel),
representing the cells of that row.