use the same validation for CSV and Excel files
Use PHPExcel to read CSV files so that we can handle the result of
calling PHPExcel::read_csv() (new) and PHPExcel::read_xls() in the
same manner. Modify processFile() so that it first reads the complete
file (either .csv or .xls) into an array and then does the checks on
missing columns or wrong IDs. This is not very performant but allows
me to quickly offer the requested ability to upload data from Excel
files without lots of changes.