reactivate document.ready-function
Take in again the body of the function that was commented out.
When functions are chained, write the dot as the last character of
a line and start the next line with the name of the chained function.
Otherwise the JavaScript validator that I use to validate the code
(http://javascriptlint.com/online_lint.php) gives a warning.