let getByVarSite() return distinct records
The SeriesCatalog may contain different records of the same SiteID,
VariableID and MethodID (e.g. differing in the QualityControlLevelID).
Without distinct() we would get duplicate records containing the
same methods. This would be confusing, e.g. when the methods are
displayed in the drop down list of the site details page.