do not use genDropLists() for fix site title
The site name was not displayed nicely. It seems that the site name
once was taken from a drop down list. Since this is not the case any
more do not use genDropLists() but use the same structure with a label.
Do not apply the class "site_title" and do not use a line break.