set global variables via helper function
The new helper function setGlobal() sets the global variable of the
given name to the given value. The change is logged to the console
and a boolean is returned indicating if the value actually changed.

The new function setGlobalDate() saves the new date in the corresponding
global variables and calls the functions to handle the update if
the date actually changed.