use grid data as input to chart
Do not use getDataAsScript() to load the data (in the form of a JavaScript
assignment to variable data_test) but read them from the jqxGrid object.
Use gridDataToSeriesData() to generate the array of data points that is
required by Highcharts.StockChart(). Use sortByFirstColumn() to sort the
data points by time which is important since otherwise the chart will not
look as expected.