improve readability of event handling
Provide get_dates() with SiteID, VariableID and MethodID in the form
of function arguments. Also pass the callback function that is called
after the dates have been looked up as argument to this function.

Split dateChangedHandler() into the part that extracts information from
the event (now in local function "handler" in document-ready function)
and the part that handles these information (setGlobalDate(), renamed to
setMinOrMaxDate()).

Split getDatesHandler() into the part that extracts information from
the ajax result (directly in callback function given to get_dates())
and the part that handles these information (setDateTimeRange()).