simplify event handling, bug fix in timeconvert()
Simplify the code by deleting the small helper function handler() and
instead defining the callback functions directly in the event binding
on('valuechanged', callback).

Fix bug in timeconvert(). The default value was not set correctly.

Use 'HH' (representing hours between 00 and 23) in the formatString
property of the date input boxes.