back to chart recreation instead of chart update
Since it is not clear how to use the update methods provided by
Highcharts to set new data, new axis labels and a new title go back
to recreating the chart with new Highcharts.StockChart().

Use updateChartConfig() to update the base configuration with the
variable settings that are passed as arguments to this function.

Extract the generation of a chart title into the new function
getTitle().

Rework the parameter lists of the functions so that it get clearer
which kind of information is used by which function.