update grid and plot if method changed
Since grid and plot were only updated if the start or end date changed
no update was performed when a new variable or method was selected for
which data are available in the exact same time range as for the
currently selected variable and method. Now, a global variable
"updateRequired" is set to true when the method changes and set back
to false after the update.