rearrange creation of popup window elements
Create the jqxDateTimeInput #date only once in initPopups() and not
always newly in editClickHandler(). There, use only the methods to
set the values of the widgets.

Remove functions saveClickHandler() and saveNewClickHandler().
Instead, define these functions, each consisting of only one call
to handleSaveClick() direcly in the binding of the 'click' event
in initPopups(). Rearrange the creation of input fields and
buttons in initPopups() so that it becomes obvious that these
elements are almost identical in the two different popup windows.

Log to tje console when the valuechanged event of the date input
fields fired.