Repair and refactor data editing, adding and deletion
editClickHandler: save the field values of the selected rows in a new
global 'editRowData'.

Use the methods 'deleterow' and 'updaterow' of jqxGrid correctly
by passing the id (in our case the ValueID) of the record instead of
the row index.

Lookup the unit (TODO: avoid repetive looking up) and call
updateChart() when the grid data has changed.

Remove dataAddOrEditHandler. Integrate its content into the new
function handleEditedOrNewValue.

Remove the local definitions of handler functions doneHandler_edit
and doneHandler_add in and handleSaveClick and remove
dataAddOrEditHandler. Instead bind an anonymous function to
ajax.done that calls handleEditedOrNewValue.

Extract the generation of the parameters object for the API calls
into the new function toApiCallParameters.