createDataPoint: get UTC from timestring if given
If the timestring in $fields['time'] ends with an indication of the
UTC offset in hours (such as in "23:00:00+01") and the UTC offset
is not given in $fields['UTCOffset'] extract the UTC offset from the
timestring and set $fields['UTCOffset'] accordingly.
By doing so we can, without changing the API interfaces, specifying
the UTCOffset when calling the datapoint/add or datapoint/edit
controller.