correct generation of timestring, add grid columns
toISOString() is returning the string of the timestamp in UTC,
however we need the string of the local time in ISO format.
Create this string manually. Append the information on the UTC
offset to the time string. In edit mode, the timestring does not
contain seconds (since the datapoint/edit method will append ":00",
TODO: make consistent with datapoint/add).

Add columns "UTCOffset" and "CensorCode" to the grid and set the
column width only for the date and time column.