allow to filter DataValues for multiple ID combinations
Instead of passing SiteID, VariableID and MethodID as different arguments
to getResultData(), pass them in an associative array $conditions.
If $conditions is an array of arrays, use each element of the array to
create a new part of the WHERE condition that filters for the
corresponding combination of IDs. Extract the code extending the field
list by columns of joined tables into a new function extendedFieldList().