allow to filter SeriesCatalog for array of SeriesIDs
In the get() method of model Sc, filter for any of the SeriesIDs
given in $seriesIDs. If $seriesIDs is not an array, create an array
with $seriesIDs being the first element. Add parameter $fields to
specify which table fields are to be returned.

In Series::getJSON(), check if the element 'VariableID' exists
before checking it for NULL. What is the sense of this check? The
SeriesCatalog should not contain records with VariableID = NULL.
Use MY_Controller::jsonEncoded() instead of json_encode().