allow export to spreadsheet for different db tables
Add endpoint "exportObjects" to the Series controller. It is called by
the jqxGrid method "exportdata" passing the filename and the content
of the grid as POST parameters. Since I did not succeed in converting
the content into an Excel file use the filename only to distinguish
which model to load and call the exportToSpreadsheet method on this
model. These changes became necessary because the default behaviour of
jqxGrid's "exportdata" method is to call a php script on the jqxGrid
server that does the file conversion. This script is not available any
more (without a developer licence).